+++ src/components/NavBar/NavBar.stories.js
@@ -8,7 +8,7 @@ import Typography from "components/Typography";
 import { Menu } from "react-feather";
 import Icon from "components/Icon";
 import Narmi from "../../assets/narmi-logo.svg";
-import { deviceBreakpoints } from "../../globalStyles";
+import { PopoverNavBar } from "../Popover/Popover.stories.js";
 
 export default {
   title: "Components/NavBar",
@@ -16,17 +16,23 @@ export default {
   decorators: [Top],
 };
 
-const Template = (args) => <NavBar {...args} />;
-const DefaultArgs = {
-  leftChildren: [],
-  rightChildren: [],
-};
+const StyledA = styled.a`
+  :hover {
+    color: var(--nds-primary-color);
+  }
+`;
 
 const PopoverButton = styled.div`
   display: flex;
   align-items: center;
   position: relative;
 `;
+
+const Template = (args) => <NavBar {...args} />;
+const DefaultArgs = {
+  leftChildren: [],
+  rightChildren: [],
+};
 
 const AccountPopover = () => {
   return (
@@ -61,12 +67,6 @@ const UserPopover = () => {
   );
 };
 
-const StyledA = styled.a`
-  :hover {
-    color: var(--nds-primary-color);
-  }
-`;
-
 const HamburgerMenuSidebar = () => {
   let icon = <Menu />;
   let menuItems = [
@@ -74,110 +74,9 @@ const HamburgerMenuSidebar = () => {
     <Typography>Documents</Typography>,
     <Typography>Suppor</Typography>,
   ];
-  menuItems = [MobileNavbar()];
+  menuItems = [PopoverNavBar()];
   return (
     <Sidebar icon={icon} slideFromRight={true} menuItems={menuItems}></Sidebar>
-  );
-};
-
-const AccountPopoverMobile = () => {
-  return (
-    <div>
-      Account
-      <Popover
-        hoverable={false}
-        label={null}
-        shiftX={"-40%"}
-        children={<NavMenu />}
-      />
-    </div>
-  );
-};
-
-const StyledOptionsList = styled.div`
-  display: flex;
-  position: absolute;
-  top: 20%;
-  left: 30%;
-  flex-direction: column;
-`;
-
-const StyledHeader = styled.span`
-  color: var(--nds-black);
-  font-weight: 600;
-  font-family: var(--nds-font-family);
-  font-size: 16px;
-  line-height: 20px;
-  @media ${`(max-width: ${deviceBreakpoints.mobileMax})`} {
-    font-size: 12px;
-  }
-`;
-
-const StyledList = styled.ul`
-  list-style-type: none;
-  padding-inline-start: 0;
-  margin-block-start: 0;
-  white-space: nowrap;
-  padding: 12px 16px;
-  margin: 0;
-
-  @media ${`(max-width: ${deviceBreakpoints.mobileMax})`} {
-    font-size: 12px;
-  }
-
-  border-right: ${(p) =>
-    p.divided && p.horizontal ? "1px solid var(--nds-grey-disabled)" : null};
-  border-bottom: ${(p) =>
-    p.divided && !p.horizontal ? "1px solid var(--nds-grey-disabled)" : null};
-
-  &:last-child {
-    border-right: none;
-  }
-`;
-
-const NavMenu = ({ divided, horizontal }) => {
-  // one way to organize Popover content, it's up to the consumer really! Popover just renders it.
-  const items = {
-    Checking: [<a>Primary Checking-1111</a>, <a>Joint Checking-2314</a>],
-    Savings: [<a>Primary Savings-2000</a>, <a>Car Savings-4232</a>],
-    Loans: [<a>Mortgage-3242</a>, <a>Auto Loan-8493</a>],
-    CDs: [<a>3 Year-8222</a>],
-  };
-
-  const els = Object.keys(items).map((header) => (
-    <StyledList divided={divided} horizontal={horizontal}>
-      <li>
-        <StyledHeader>{header}</StyledHeader>
-      </li>
-      {items[header].map((c) => (
-        <li key={c}>{c}</li>
-      ))}
-    </StyledList>
-  ));
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        flexFlow: horizontal ? "row nowrap" : "column nowrap",
-      }}
-    >
-      {els}
-    </div>
-  );
-};
-
-const MobileNavbar = () => {
-  let options = [
-    <AccountPopoverMobile />,
-    <p>Documents</p>,
-    <p>Support</p>,
-    <p>Tools</p>,
-  ];
-  return (
-    <div>
-      <StyledOptionsList>{options}</StyledOptionsList>
-    </div>
   );
 };
