+++ src/components/Popover/Popover.stories.js
@@ -19,7 +19,7 @@ const StyledHeader = styled.span`
   line-height: 20px;
   @media ${`(max-width: ${deviceBreakpoints.mobileMax})`} {
     font-size: 12px;
-}
+  }
 `;
 const StyledList = styled.ul`
   list-style-type: none;
@@ -28,9 +28,9 @@ const StyledList = styled.ul`
   white-space: nowrap;
   padding: 12px 16px;
   margin: 0;
-  
+
   @media ${`(max-width: ${deviceBreakpoints.mobileMax})`} {
-      font-size: 12px;
+    font-size: 12px;
   }
 
   border-right: ${(p) =>
@@ -42,7 +42,7 @@ const StyledList = styled.ul`
     border-right: none;
   }
 `;
-const NavMenu = ({ divided, horizontal }) => {
+export const NavMenu = ({ divided, horizontal }) => {
   // one way to organize Popover content, it's up to the consumer really! Popover just renders it.
   const items = {
     Checking: [<a>Primary Checking-1111</a>, <a>Joint Checking-2314</a>],
@@ -52,9 +52,9 @@ const NavMenu = ({ divided, horizontal }) => {
   };
 
   const els = Object.keys(items).map((header) => (
-    <StyledList divided={divided} horizontal={horizontal} >
+    <StyledList divided={divided} horizontal={horizontal}>
       <li>
-        <StyledHeader >{header}</StyledHeader>
+        <StyledHeader>{header}</StyledHeader>
       </li>
       {items[header].map((c) => (
         <li key={c}>{c}</li>
@@ -128,9 +128,8 @@ OnHoverLabel.args = {
   children: <NavMenu divided horizontal />,
 };
 
-
-const StyledOptionsList = styled.div`
-  display: flex; 
+export const StyledOptionsList = styled.div`
+  display: flex;
   position: absolute;
   top: 20%;
   left: 30%;
@@ -139,24 +138,28 @@ const StyledOptionsList = styled.div`
 
 const AccountPopover = () => {
   return (
-      <div>Account 
-          <Popover 
-            hoverable={false}
-            label={null}
-            shiftX={"-40%"}
-            children={<NavMenu />}
-          />
-      </div>
+    <div>
+      Account
+      <Popover
+        hoverable={false}
+        label={null}
+        shiftX={"-40%"}
+        children={<NavMenu />}
+      />
+    </div>
   );
 };
 
-export const InNavbar = () => {
-  let options = [<AccountPopover />, <p>Documents</p>, <p>Support</p>, <p>Tools</p>];
-  return (    
-      <div>
-        <StyledOptionsList>
-          {options}
-        </StyledOptionsList>
-      </div>
+export const PopoverNavBar = () => {
+  let options = [
+    <AccountPopover />,
+    <p>Documents</p>,
+    <p>Support</p>,
+    <p>Tools</p>,
+  ];
+  return (
+    <div>
+      <StyledOptionsList>{options}</StyledOptionsList>
+    </div>
   );
 };
