+++ src/index.js
@@ -3,7 +3,6 @@ const ButtonGroup = require("./components/ButtonGroup").default;
 const Input = require("./components/Input").default;
 const PlainButton = require("./components/PlainButton").default;
 const Popover = require("./components/Popover").default;
-const Sidebar = require("./components/Sidebar").default;
 const NavBar = require("./components/NavBar").default;
 const GlobalStyles = require("./globalStyles").default;
 
@@ -13,7 +12,6 @@ module.exports = {
   Input,
   PlainButton,
   Popover,
-  Sidebar,
   NavBar,
   GlobalStyles,
 };
